Updated test func-declarations
